+++ projects/react/ImageEditorWrapper.js
@@ -3,7 +3,7 @@ import ImageEditor from './ImageEditor';
 import { Container } from './styledComponents';
 import { ThemeProvider } from 'styled-components';
 import { Modal } from './components/Modal';
-import { CLOUDIMAGE_OPERATIONS, TOOLS, UPLOADER, ON_CLOSE_STATUSES } from './config';
+import { CLOUDIMAGE_OPERATIONS, TOOLS, UPLOADER, ON_CLOSE_STATUSES, STANDARD_FONTS } from './config';
 import './assets/fonts/filerobot-font.css';
 import translations from './assets/i18n';
 import dark from './assets/theme/dark';
@@ -21,7 +21,7 @@ class ImageEditorWrapper extends Component {
     config.language = (config.translations[config.language] || translations[config.language]) ? config.language : 'en';
     config.theme = config.theme || {};
     config.theme.colors = config.theme.colors || {};
-    config.theme.fonts = config.theme.fonts || {};
+    config.theme.fonts = config.theme.fonts || STANDARD_FONTS;
     config.colorScheme = config.colorScheme || 'dark';
     config.platform = config.platform || 'filerobot';
 
@@ -39,10 +39,7 @@ class ImageEditorWrapper extends Component {
           ...(config.colorScheme === 'light' ? light : dark).colors,
           ...config.theme.colors
         },
-        fonts: {
-          ...(config.colorScheme === 'light' ? light : dark).fonts,
-          ...config.theme.fonts
-        }
+        fonts: config.theme.fonts
       }
     }
   }